Add tests for single reward happy path
Taking the TDD approach I have written the tests to check
that the RewardsService returns correct rewards if the
customer is eligible. I had to refactor the tests to use
async since the EligibilityService will need to use
promises if the communication between the two services will
be done over http of other asynchronous means.